Spectra v2.0.0 Release

+++ classes/class-uagb-admin-helper.php
@@ -109,81 +109,6 @@ if ( ! class_exists( 'UAGB_Admin_Helper' ) ) {
 		}
 
 		/**
-		 * Is Knowledgebase.
-		 *
-		 * @return string
-		 * @since 0.0.1
-		 */
-		public static function knowledgebase_data() {
-
-			$knowledgebase = array(
-				'enable_knowledgebase' => true,
-				'knowledgebase_url'    => 'https://www.ultimategutenberg.com/docs/?utm_source=uag-dashboard&utm_medium=link&utm_campaign=uag-dashboard',
-			);
-
-			return $knowledgebase;
-		}
-
-		/**
-		 * Is Knowledgebase.
-		 *
-		 * @return string
-		 * @since 0.0.1
-		 */
-		public static function support_data() {
-
-			$support = array(
-				'enable_support' => true,
-				'support_url'    => 'https://www.ultimategutenberg.com/support/?utm_source=uag-dashboard&utm_medium=link&utm_campaign=uag-dashboard',
-			);
-
-			return $support;
-		}
-
-		/**
-		 * Get flag if more than 5 pages are build using UAG.
-		 *
-		 * @since  1.10.0
-		 * @return boolean true/false Flag if more than 5 pages are build using UAG.
-		 */
-		public static function show_rating_notice() {
-
-			$posts_created_with_uag = get_option( 'posts-created-with-uagb' );
-
-			if ( false === $posts_created_with_uag ) {
-				$query_args = array(
-					'posts_per_page' => 100,
-					'post_status'    => 'publish',
-					'post_type'      => 'any',
-				);
-
-				$query = new WP_Query( $query_args );
-
-				$uag_post_count = 0;
-
-				if ( isset( $query->post_count ) && $query->post_count > 0 ) {
-					foreach ( $query->posts as $key => $post ) {
-						if ( $uag_post_count >= 5 ) {
-							break;
-						}
-
-						if ( false !== strpos( $post->post_content, '<!-- wp:uagb/' ) ) {
-							$uag_post_count++;
-						}
-					}
-				}
-
-				if ( $uag_post_count >= 5 ) {
-					update_option( 'posts-created-with-uagb', $uag_post_count );
-
-					$posts_created_with_uag = $uag_post_count;
-				}
-			}
-
-			return ( $posts_created_with_uag >= 5 );
-		}
-
-		/**
 		 *  Get Specific Stylesheet
 		 *
 		 * @since 1.13.4
@@ -198,6 +123,7 @@ if ( ! class_exists( 'UAGB_Admin_Helper' ) ) {
 			$is_already_icon_list = false;
 			$is_already_button    = false;
 			$is_already_faq       = false;
+
 			foreach ( UAGB_Config::$block_attributes as $key => $block ) {
 
 				$block_name = str_replace( 'uagb/', '', $key );
@@ -275,18 +201,29 @@ if ( ! class_exists( 'UAGB_Admin_Helper' ) ) {
 				}
 			}
 
+			// Load common CSS for all the blocks.
+			$combined[] = 'extensions';
+
 			$wp_upload_dir = UAGB_Helper::get_uag_upload_dir_path();
 			$combined_path = $wp_upload_dir . 'custom-style-blocks.css';
-			wp_delete_file( $combined_path );
+
+			if ( file_exists( $combined_path ) ) {
+				wp_delete_file( $combined_path );
+			}
 
 			$style = '';
 
 			$wp_filesystem = uagb_filesystem();
 
 			foreach ( $combined as $key => $c_block ) {
-				$style .= $wp_filesystem->get_contents( UAGB_DIR . 'assets/css/blocks/' . $c_block . '.css' );
 
+				$style_file = UAGB_DIR . 'assets/css/blocks/' . $c_block . '.css';
+
+				if ( file_exists( $style_file ) ) {
+					$style .= $wp_filesystem->get_contents( $style_file );
+				}
 			}
+
 			$wp_filesystem->put_contents( $combined_path, $style, FS_CHMOD_FILE );
 		}
 
@@ -301,7 +238,7 @@ if ( ! class_exists( 'UAGB_Admin_Helper' ) ) {
 
 			$rollback_versions = get_transient( 'uag_rollback_versions_' . UAGB_VER );
 
-			if ( ! $rollback_versions || empty( $rollback_versions ) ) {
+			if ( empty( $rollback_versions ) ) {
 
 				$max_versions = 10;
 
@@ -371,6 +308,36 @@ if ( ! class_exists( 'UAGB_Admin_Helper' ) ) {
 			return 1;
 		}
 
+		/**
+		 * Get Global Content Width
+		 *
+		 * @since 2.0.0
+		 * @return int
+		 * @access public
+		 */
+		public static function get_global_content_width() {
+
+			$content_width             = self::get_admin_settings_option( 'uag_content_width', '' );
+			$content_width_third_party = apply_filters( 'spectra_global_content_width', 'default' );
+			$astra_content_width       = false;
+			self::update_admin_settings_option( 'uag_content_width_set_by', __( 'Spectra', 'ultimate-addons-for-gutenberg' ) );
+			if ( function_exists( 'astra_get_option' ) ) {
+				$astra_content_width = astra_get_option( 'site-content-width' );
+			}
+
+			if ( '' === $content_width ) {
+				if ( $astra_content_width ) {
+					$content_width = intval( $astra_content_width );
+					self::update_admin_settings_option( 'uag_content_width_set_by', __( 'Astra Theme', 'ultimate-addons-for-gutenberg' ) );
+				}
+				if ( 'default' !== $content_width_third_party ) {
+					$content_width = intval( $content_width_third_party );
+					self::update_admin_settings_option( 'uag_content_width_set_by', __( 'Filter added through any 3rd Party Theme/Plugin.', 'ultimate-addons-for-gutenberg' ) );
+				}
+			}
+
+			return $content_width;
+		}
 	}
 
 	/**
